fix(coding-agent): use focused eval harness
