+++ biome.json
@@ -34,7 +34,6 @@
 			"!**/test-sessions.ts",
 			"!**/models.generated.ts",
 			"!**/*.models.ts",
-			"!packages/evals/src/pi-harness.ts",
 			"!packages/mom/data/**/*",
 			"!!**/node_modules"
 		]
