+++ packages/evals/package.json
@@ -9,7 +9,6 @@
 	},
 	"devDependencies": {
 		"@earendil-works/pi-coding-agent": "^0.82.1",
-		"@earendil-works/pi-ai": "^0.82.1",
 		"@types/node": "24.12.4",
 		"shx": "0.4.0",
 		"typescript": "5.9.3",
